refactor: move root.go to internal/cli and remove business package
- Move cmd/root.go to internal/cli/root.go (root command is a CLI concern)
- Remove business/ package entirely
- Move business logic back to cmd/ package as per original design
- Update all imports: business.* → cmd.*
- Remove cli import from cmd/test_helpers.go (inline formatter)
- Fix circular dependency: main → internal/cli → cmd (one-way flow)

This matches the original architecture design:
- cmd/ = pure business logic functions
- internal/cli/ = all Cobra-related code (root + wrappers)
- No more business/ package
