+++ cmd/tidy_test_logic.go
@@ -1,4 +1,4 @@
-package business
+package cmd
 
 import (
 	"bytes"
